+++ lib/models/institution_info.dart
@@ -1,0 +1,15 @@
+class InstitutionInfo {
+  final String name;
+  final String idIcon;
+  final String primarySource;
+  final String officialSite;
+  final String primarySourceLink;
+
+  InstitutionInfo({
+    required this.name,
+    required this.idIcon,
+    required this.primarySource,
+    required this.officialSite,
+    required this.primarySourceLink,
+  });
+}